Add git support to the zsh prompt
Use the zsh vcs_info module to create an information string when in git
repositories. At present, the string consists of branch, hash, and file
tracking status. This is in the git/prompt.zsh file, which at present
must be manually sourced and run on shell startup (see zsh/user/zshrc.sh
line 39).

TODO: Set up the sysconfig script to be able to install shell config
files for both system-wide and per-user storage, and configure the shell
rc files to load them.

TODO: Make the vcs_info callback functions work, so we can see untracked
files as well as staged and modified. Look into having vcs_info provide
the directory string as well. This will require a custom function to
swap "${HOME}" for "~", as vcs_info's %R variable does not do so for us.